feat(app): add async lock
Ensure that main keeps running in background.
This will help if a container goes down and needs
to push the interface back in.

This also means that the intial config needs to
be updated everytime a new REST API call is done.
This will allow the background runner to pick up
the next configuration and continue building on top
of it.

Change-Id: I165d7d1bbf90f32f7f3c42f45ae434e8ad3a188f
Signed-off-by: Ketan Tewari <ktewari@libertyglobal.com>
Reviewed-on: http://10.64.38.13:8080/gerrit/c/github/raikou-net/+/9911
Reviewed-by: Ketan Tewari <ktewari.contractor@libertyglobal.com>
Tested-by: Ketan Tewari <ktewari.contractor@libertyglobal.com>